… go.sum
… vendor/cloud.google.com/go/compute/internal/version.go
… vendor/github.com/asaskevich/govalidator/validator.go
… vendor/github.com/aws/aws-sdk-go-v2/CHANGELOG.md
… vendor/github.com/aws/aws-sdk-go-v2/CODE_OF_CONDUCT.md
… vendor/github.com/aws/aws-sdk-go-v2/CONTRIBUTING.md
… vendor/github.com/aws/aws-sdk-go-v2/DESIGN.md
… vendor/github.com/aws/aws-sdk-go-v2/Makefile
… vendor/github.com/aws/aws-sdk-go-v2/NOTICE.txt
… vendor/github.com/aws/aws-sdk-go-v2/README.md
… vendor/github.com/aws/aws-sdk-go-v2/aws/go_module_metadata.go
… vendor/github.com/aws/aws-sdk-go-v2/aws/protocol/query/array.go
… vendor/github.com/aws/aws-sdk-go-v2/aws/protocol/xml/error_utils.go
… vendor/github.com/aws/aws-sdk-go-v2/aws/ratelimit/token_rate_limit.go
… vendor/github.com/aws/aws-sdk-go-v2/aws/retry/middleware.go
… vendor/github.com/aws/aws-sdk-go-v2/aws/signer/internal/v4/headers.go
… vendor/github.com/aws/aws-sdk-go-v2/config/CHANGELOG.md
… vendor/github.com/aws/aws-sdk-go-v2/config/go_module_metadata.go
… vendor/github.com/aws/aws-sdk-go-v2/config/resolve_credentials.go
… vendor/github.com/aws/aws-sdk-go-v2/credentials/CHANGELOG.md
… vendor/github.com/aws/aws-sdk-go-v2/credentials/ec2rolecreds/doc.go
… vendor/github.com/aws/aws-sdk-go-v2/credentials/go_module_metadata.go
… vendor/github.com/aws/aws-sdk-go-v2/credentials/processcreds/provider.go
… vendor/github.com/aws/aws-sdk-go-v2/feature/ec2/imds/CHANGELOG.md
… vendor/github.com/aws/aws-sdk-go-v2/feature/ec2/imds/api_client.go
… vendor/github.com/aws/aws-sdk-go-v2/feature/ec2/imds/go_module_metadata.go
… vendor/github.com/aws/aws-sdk-go-v2/feature/ec2/imds/token_provider.go
… vendor/github.com/aws/aws-sdk-go-v2/internal/configsources/CHANGELOG.md
… vendor/github.com/aws/aws-sdk-go-v2/internal/configsources/go_module_metadata.go
… vendor/github.com/aws/aws-sdk-go-v2/internal/endpoints/v2/CHANGELOG.md